Основная ViewModel разделена на части
