+++ src/RevitCopyInteriorSpecs/ViewModels/MainViewModel.cs
@@ -2,7 +2,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Linq;
 using System.Windows;
 using System.Windows.Input;
@@ -16,6 +15,7 @@ using dosymep.WPF.Commands;
 using dosymep.WPF.ViewModels;
 
 using RevitCopyInteriorSpecs.Models;
+using RevitCopyInteriorSpecs.ViewModels.MainViewModelParts;
 
 namespace RevitCopyInteriorSpecs.ViewModels {
     internal class MainViewModel : BaseViewModel {
@@ -26,27 +26,8 @@ namespace RevitCopyInteriorSpecs.ViewModels {
         private string _errorText;
         private string _saveProperty;
 
-        private List<ViewSchedule> _selectedSpecs;
-        private List<Level> _levels;
-        private List<Phase> _phases;
-
-        private ObservableCollection<TaskInfoVM> _tasksForWork = new ObservableCollection<TaskInfoVM>();
-        private TaskInfoVM _selectedTask;
-
-        private string _groupTypeParamName = "ФОП_Тип квартиры";
-        private string _levelShortNameParamName = "ФОП_Этаж";
-        private string _firstDispatcherGroupingLevelParamName = "_Стадия Проекта";
-        private string _secondDispatcherGroupingLevelParamName = "_Группа Видов";
-        private string _thirdDispatcherGroupingLevelParamName = "Назначение вида";
-        private string _generalGroupType;
-        private Level _generalLevel;
-        private string _generalLevelShortName;
-        private Phase _generalPhase;
-        private string _generalFirstDispatcherGroupingLevel;
-        private string _generalSecondDispatcherGroupingLevel;
-        private string _generalThirdDispatcherGroupingLevel;
-
         private string _errorReport = string.Empty;
+
 
         public MainViewModel(
             PluginConfig pluginConfig,
@@ -60,33 +41,19 @@ namespace RevitCopyInteriorSpecs.ViewModels {
             LoadViewCommand = RelayCommand.Create(LoadView);
             AcceptViewCommand = RelayCommand.Create(AcceptView, CanAcceptView);
 
-            AddTaskCommand = RelayCommand.Create(AddTask);
-            DeleteTaskCommand = RelayCommand.Create(DeleteTask, CanDeleteTask);
-
-            GeneralGroupTypeChangedCommand = RelayCommand.Create(GeneralGroupTypeChanged);
-            GeneralLevelChangedCommand = RelayCommand.Create(GeneralLevelChanged);
-            GeneralLevelShortNameChangedCommand = RelayCommand.Create(GeneralLevelShortNameChanged);
-            GeneralPhaseChangedCommand = RelayCommand.Create(GeneralPhaseChanged);
-            GeneralFirstDispatcherGroupingLevelChangedCommand = RelayCommand.Create(GeneralFirstDispatcherGroupingLevelChanged);
-            GeneralSecondDispatcherGroupingLevelChangedCommand = RelayCommand.Create(GeneralSecondDispatcherGroupingLevelChanged);
-            GeneralThirdDispatcherGroupingLevelChangedCommand = RelayCommand.Create(GeneralThirdDispatcherGroupingLevelChanged);
+            SelectedSpecsVM = new SelectedSpecsViewModel(_revitRepository);
+            TasksVM = new TasksViewModel(_revitRepository);
+            ParametersVM = new ParametersViewModel();
         }
+
+
+        public SelectedSpecsViewModel SelectedSpecsVM { get; }
+        public TasksViewModel TasksVM { get; }
+        public ParametersViewModel ParametersVM { get; }
 
 
         public ICommand LoadViewCommand { get; }
         public ICommand AcceptViewCommand { get; }
-
-        public ICommand AddTaskCommand { get; }
-        public ICommand DeleteTaskCommand { get; }
-
-        public ICommand GeneralGroupTypeChangedCommand { get; }
-        public ICommand GeneralLevelChangedCommand { get; }
-        public ICommand GeneralLevelShortNameChangedCommand { get; }
-        public ICommand GeneralPhaseChangedCommand { get; }
-        public ICommand GeneralFirstDispatcherGroupingLevelChangedCommand { get; }
-        public ICommand GeneralSecondDispatcherGroupingLevelChangedCommand { get; }
-        public ICommand GeneralThirdDispatcherGroupingLevelChangedCommand { get; }
-
 
 
         public string ErrorText {
@@ -99,112 +66,21 @@ namespace RevitCopyInteriorSpecs.ViewModels {
             set => this.RaiseAndSetIfChanged(ref _saveProperty, value);
         }
 
-        public List<ViewSchedule> SelectedSpecs {
-            get => _selectedSpecs;
-            set => this.RaiseAndSetIfChanged(ref _selectedSpecs, value);
-        }
-
-        public List<Level> Levels {
-            get => _levels;
-            set => this.RaiseAndSetIfChanged(ref _levels, value);
-        }
-
-        public List<Phase> Phases {
-            get => _phases;
-            set => this.RaiseAndSetIfChanged(ref _phases, value);
-        }
-
-        public ObservableCollection<TaskInfoVM> TasksForWork {
-            get => _tasksForWork;
-            set => this.RaiseAndSetIfChanged(ref _tasksForWork, value);
-        }
-
-        public TaskInfoVM SelectedTask {
-            get => _selectedTask;
-            set => this.RaiseAndSetIfChanged(ref _selectedTask, value);
-        }
-
-
-        public string GroupTypeParamName {
-            get => _groupTypeParamName;
-            set => this.RaiseAndSetIfChanged(ref _groupTypeParamName, value);
-        }
-
-        public string LevelShortNameParamName {
-            get => _levelShortNameParamName;
-            set => this.RaiseAndSetIfChanged(ref _levelShortNameParamName, value);
-        }
-
-        public string FirstDispatcherGroupingLevelParamName {
-            get => _firstDispatcherGroupingLevelParamName;
-            set => this.RaiseAndSetIfChanged(ref _firstDispatcherGroupingLevelParamName, value);
-        }
-
-        public string SecondDispatcherGroupingLevelParamName {
-            get => _secondDispatcherGroupingLevelParamName;
-            set => this.RaiseAndSetIfChanged(ref _secondDispatcherGroupingLevelParamName, value);
-        }
-
-        public string ThirdDispatcherGroupingLevelParamName {
-            get => _thirdDispatcherGroupingLevelParamName;
-            set => this.RaiseAndSetIfChanged(ref _thirdDispatcherGroupingLevelParamName, value);
-        }
-
-
-        public string GeneralGroupType {
-            get => _generalGroupType;
-            set => this.RaiseAndSetIfChanged(ref _generalGroupType, value);
-        }
-
-        public Level GeneralLevel {
-            get => _generalLevel;
-            set => this.RaiseAndSetIfChanged(ref _generalLevel, value);
-        }
-
-        public string GeneralLevelShortName {
-            get => _generalLevelShortName;
-            set => this.RaiseAndSetIfChanged(ref _generalLevelShortName, value);
-        }
-
-        public Phase GeneralPhase {
-            get => _generalPhase;
-            set => this.RaiseAndSetIfChanged(ref _generalPhase, value);
-        }
-
-        public string GeneralFirstDispatcherGroupingLevel {
-            get => _generalFirstDispatcherGroupingLevel;
-            set => this.RaiseAndSetIfChanged(ref _generalFirstDispatcherGroupingLevel, value);
-        }
-
-        public string GeneralSecondDispatcherGroupingLevel {
-            get => _generalSecondDispatcherGroupingLevel;
-            set => this.RaiseAndSetIfChanged(ref _generalSecondDispatcherGroupingLevel, value);
-        }
-
-        public string GeneralThirdDispatcherGroupingLevel {
-            get => _generalThirdDispatcherGroupingLevel;
-            set => this.RaiseAndSetIfChanged(ref _generalThirdDispatcherGroupingLevel, value);
-        }
-
-
 
         private void LoadView() {
             LoadConfig();
 
-            SelectedSpecs = _revitRepository.GetSelectedSpecs();
-            Levels = _revitRepository.GetElements<Level>();
-            Phases = _revitRepository.GetElements<Phase>();
-
-            TasksForWork.Add(new TaskInfoVM());
+            TasksVM.TasksForWork.Add(new TaskInfoVM());
         }
+
 
         private void AcceptView() {
             SaveConfig();
 
             using(Transaction transaction = _revitRepository.Document.StartTransaction("Копирование спецификаций АИ")) {
 
-                foreach(TaskInfoVM task in TasksForWork) {
-                    foreach(ViewSchedule spec in SelectedSpecs) {
+                foreach(TaskInfoVM task in TasksVM.TasksForWork) {
+                    foreach(ViewSchedule spec in SelectedSpecsVM.SelectedSpecs) {
                         string specOldName = spec.Name;
 
                         string newViewSpecName = $"{specOldName}_{task.Phase.Name}_{task.GroupType}_{task.LevelShortName}";
@@ -217,9 +93,9 @@ namespace RevitCopyInteriorSpecs.ViewModels {
                         }
 
                         DispatcherOption dispatcherOption = new DispatcherOption() {
-                            FirstGroupingLevelParamName = FirstDispatcherGroupingLevelParamName,
-                            SecondGroupingLevelParamName = SecondDispatcherGroupingLevelParamName,
-                            ThirdGroupingLevelParamName = ThirdDispatcherGroupingLevelParamName,
+                            FirstGroupingLevelParamName = ParametersVM.FirstDispatcherGroupingLevelParamName,
+                            SecondGroupingLevelParamName = ParametersVM.SecondDispatcherGroupingLevelParamName,
+                            ThirdGroupingLevelParamName = ParametersVM.ThirdDispatcherGroupingLevelParamName,
 
                             FirstGroupingLevelParamValue = task.FirstDispatcherGroupingLevel,
                             SecondGroupingLevelParamValue = task.SecondDispatcherGroupingLevel,
@@ -228,9 +104,9 @@ namespace RevitCopyInteriorSpecs.ViewModels {
 
                         SetSpecParams(newViewSpec, dispatcherOption);
 
-                        ChangeSpecFilters(newViewSpec, GroupTypeParamName, task.GroupType);
+                        ChangeSpecFilters(newViewSpec, ParametersVM.GroupTypeParamName, task.GroupType);
                         ChangeSpecFilters(newViewSpec, "Уровень", task.Level.Id);
-                        ChangeSpecFilters(newViewSpec, LevelShortNameParamName, task.LevelShortName);
+                        ChangeSpecFilters(newViewSpec, ParametersVM.LevelShortNameParamName, task.LevelShortName);
                     }
                 }
 
@@ -245,12 +121,12 @@ namespace RevitCopyInteriorSpecs.ViewModels {
 
         private bool CanAcceptView() {
 
-            if(TasksForWork.Count == 0) {
+            if(TasksVM.TasksForWork.Count == 0) {
                 ErrorText = _localizationService.GetLocalizedString("MainWindow.NoTasksCreated");
                 return false;
             }
 
-            foreach(TaskInfoVM task in TasksForWork) {
+            foreach(TaskInfoVM task in TasksVM.TasksForWork) {
                 if(task.Level is null) {
                     ErrorText = _localizationService.GetLocalizedString("MainWindow.NotAllLevelsSelected");
                     return false;
@@ -265,6 +141,7 @@ namespace RevitCopyInteriorSpecs.ViewModels {
             ErrorText = string.Empty;
             return true;
         }
+
 
         private void LoadConfig() {
             RevitSettings setting = _pluginConfig.GetSettings(_revitRepository.Document);
@@ -353,65 +230,6 @@ namespace RevitCopyInteriorSpecs.ViewModels {
             }
 
             specificationDefinition.SetFilters(newScheduleFilters);
-        }
-
-
-
-        /// <summary>
-        /// Добавляет задачу в список. 
-        /// Задача содержит информацию о начальном и конечном уровне, с которыми нужно работать; выбранную область видимости и спеки
-        /// </summary>
-        private void AddTask() {
-            TasksForWork.Add(new TaskInfoVM());
-        }
-
-        /// <summary>
-        /// Удаляет выбранную в интерфейсе задачу из списка. 
-        /// </summary>
-        private void DeleteTask() {
-            TasksForWork.Remove(SelectedTask);
-        }
-
-        private bool CanDeleteTask() {
-            if(SelectedTask is null) {
-                return false;
-            }
-            return true;
-        }
-
-
-        private void GeneralGroupTypeChanged() {
-            SetTasksPropValues("GroupType", GeneralGroupType);
-        }
-
-        private void GeneralLevelChanged() {
-            SetTasksPropValues("Level", GeneralLevel);
-        }
-
-        private void GeneralLevelShortNameChanged() {
-            SetTasksPropValues("LevelShortName", GeneralLevelShortName);
-        }
-
-        private void GeneralPhaseChanged() {
-            SetTasksPropValues("Phase", GeneralPhase);
-        }
-
-        private void GeneralFirstDispatcherGroupingLevelChanged() {
-            SetTasksPropValues("FirstDispatcherGroupingLevel", GeneralFirstDispatcherGroupingLevel);
-        }
-
-        private void GeneralSecondDispatcherGroupingLevelChanged() {
-            SetTasksPropValues("SecondDispatcherGroupingLevel", GeneralSecondDispatcherGroupingLevel);
-        }
-        private void GeneralThirdDispatcherGroupingLevelChanged() {
-            SetTasksPropValues("ThirdDispatcherGroupingLevel", GeneralThirdDispatcherGroupingLevel);
-        }
-
-        private void SetTasksPropValues(string propName, object propValue) {
-            foreach(TaskInfoVM task in TasksForWork) {
-                var prop = task.GetType().GetProperty(propName);
-                prop.SetValue(task, propValue);
-            }
         }
     }
 }
